Added heart UI plus way to customize weapons

+++ Assets/Scripts/Entities/Health.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Health : MonoBehaviour {
 
-    public int health = 3;
+    public float health = 3;
     public bool dead = false;
     public float hitTimer = 0;
     public float hitTimerDelay = 1f;
@@ -23,7 +23,7 @@ public class Health : MonoBehaviour {
         dead = health <= 0;
     }
 
-    public void Hurt(int damage = 1)
+    public void Hurt(float damage = 1)
     {
         if (hitTimer <= 0)
         {
